Now API.hs uses conduits to decode the incoming JSON message.
In theory this could make things faster by only loading and processing the
result a chunk at a time, but the size of the results returned from the
server are small enough that it probably doesn't make a difference.  The
main benefit is just that I wanted to get more experience with using
conduits in my code.