Greatly simplified everything by getting rid of the paragraph tree.
Now the text of the story is passed directly to the client instead of
having to be parsed into an intermediate data structure and then rendered
back to HTML, which lets me delete a ton of code and also lets me use
arbitrary HTML without having to extend my data structure each time.

Woohoo!