Now not founds are no longer redirected to /
I found out the hard way while working on another patch that it is not a
good idea to have paths that do not exist be redirected to paths that do
exist.  The original intent was that it would help the user if they mistype
a URL but this is unlikely to ever come up whereas the browser requesting
something that doesn't exist -- namely, the site icon -- happens all of the
time.