Experimented with just creating an Outcomes directly instead of using TH.
I think that this is in retrospect a better way of creating Outcomes
because it is a bit easier to read, a bit more compact since some sections
fit on a single line, allows for sharing of stories when, for example, two
branches have the same ending, and better error reporting since GHC checks
that all of the fields are present instead of having to hear from the
Template Haskell code when expected sections are abasent.  Also, the
constructor makes the kind of Outcomes explicit, rather than having to
infer them from the section headers that are present.